fix: address review comments — processPayment, onAction propagation, stories
