+++ packages/lib/src/core/core.ts
@@ -399,9 +399,8 @@ class Core implements ICore {
                 }
                 const order = response.order;
                 if (order && (order.remainingAmount?.value ?? 0) > 0) {
-                    return this.update({ order }).then(() => {
-                        this.options.onOrderUpdated?.({ order });
-                    });
+                    this.options.onOrderUpdated?.({ order });
+                    return;
                 }
                 cleanupFinalResult(response);
                 this.options.onPaymentCompleted?.(response);
